move standalone example into its own dir
